Refactor webhook, To #34515520
